compiling with gateway
