+++ crates/rattler-bin/Cargo.toml
@@ -30,7 +30,7 @@ once_cell = { workspace = true }
 rattler = { path="../rattler", version = "0.32.1", default-features = false, features = ["indicatif"] }
 rattler_conda_types = { path="../rattler_conda_types", version = "0.31.0", default-features = false }
 rattler_networking = { path="../rattler_networking", version = "0.22.5", default-features = false, features = ["gcs", "s3"] }
-rattler_repodata_gateway = { path="../rattler_repodata_gateway", version = "0.21.38", default-features = false, features = ["gateway"] }
+rattler_repodata_gateway = { path="../rattler_repodata_gateway", version = "0.21.38", default-features = false, features = ["gateway", "memmap2"] }
 rattler_solve = { path="../rattler_solve", version = "1.3.9", default-features = false, features = ["resolvo", "libsolv_c"] }
 rattler_virtual_packages = { path="../rattler_virtual_packages", version = "2.0.4", default-features = false }
 rattler_cache = { path="../rattler_cache", version = "0.3.10", default-features = false }
