+++ crates/rattler_repodata_gateway/src/gateway/builder.rs
@@ -1,10 +1,10 @@
 use crate::gateway::GatewayInner;
 use crate::{ChannelConfig, Gateway};
 use dashmap::DashMap;
+#[cfg(not(target_arch = "wasm32"))]
 use rattler_cache::package_cache::PackageCache;
 use reqwest::Client;
 use reqwest_middleware::ClientWithMiddleware;
-use std::path::PathBuf;
 use std::sync::Arc;
 
 /// A builder for constructing a [`Gateway`].
@@ -12,7 +12,9 @@ use std::sync::Arc;
 pub struct GatewayBuilder {
     channel_config: ChannelConfig,
     client: Option<ClientWithMiddleware>,
-    cache: Option<PathBuf>,
+    #[cfg(not(target_arch = "wasm32"))]
+    cache: Option<std::path::PathBuf>,
+    #[cfg(not(target_arch = "wasm32"))]
     package_cache: Option<PackageCache>,
     max_concurrent_requests: Option<usize>,
 }
@@ -50,25 +52,29 @@ impl GatewayBuilder {
     }
 
     /// Set the directory to use for caching repodata.
+    #[cfg(not(target_arch = "wasm32"))]
     #[must_use]
-    pub fn with_cache_dir(mut self, cache: impl Into<PathBuf>) -> Self {
+    pub fn with_cache_dir(mut self, cache: impl Into<std::path::PathBuf>) -> Self {
         self.set_cache_dir(cache);
         self
     }
 
     /// Add package cache to the builder to store packages.
+    #[cfg(not(target_arch = "wasm32"))]
     pub fn with_package_cache(mut self, package_cache: PackageCache) -> Self {
         self.set_package_cache(package_cache);
         self
     }
 
     /// Set the directory to use for caching repodata.
-    pub fn set_cache_dir(&mut self, cache: impl Into<PathBuf>) -> &mut Self {
+    #[cfg(not(target_arch = "wasm32"))]
+    pub fn set_cache_dir(&mut self, cache: impl Into<std::path::PathBuf>) -> &mut Self {
         self.cache = Some(cache.into());
         self
     }
 
     /// Set the directory to use for caching packages.
+    #[cfg(not(target_arch = "wasm32"))]
     pub fn set_package_cache(&mut self, package_cache: PackageCache) -> &mut Self {
         self.package_cache = Some(package_cache);
         self
@@ -93,12 +99,14 @@ impl GatewayBuilder {
             .client
             .unwrap_or_else(|| ClientWithMiddleware::from(Client::new()));
 
+        #[cfg(not(target_arch = "wasm32"))]
         let cache = self.cache.unwrap_or_else(|| {
             dirs::cache_dir()
-                .unwrap_or_else(|| PathBuf::from("."))
+                .unwrap_or_else(|| std::path::PathBuf::from("."))
                 .join("rattler/cache")
         });
 
+        #[cfg(not(target_arch = "wasm32"))]
         let package_cache = self.package_cache.unwrap_or(PackageCache::new(
             cache.join(rattler_cache::PACKAGE_CACHE_DIR),
         ));
@@ -109,7 +117,9 @@ impl GatewayBuilder {
                 subdirs: DashMap::default(),
                 client,
                 channel_config: self.channel_config,
+                #[cfg(not(target_arch = "wasm32"))]
                 cache,
+                #[cfg(not(target_arch = "wasm32"))]
                 package_cache,
                 concurrent_requests_semaphore: Arc::new(tokio::sync::Semaphore::new(
                     max_concurrent_requests,
