+++ crates/rattler_repodata_gateway/src/gateway/channel_config.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use rattler_conda_types::ChannelUrl;
 use url::Url;
 
-use crate::fetch::with_cache::CacheAction;
+use crate::fetch::CacheAction;
 
 /// Describes additional properties that influence how the gateway fetches
 /// repodata for a specific channel.
