+++ crates/rattler_repodata_gateway/src/gateway/mod.rs
@@ -1,6 +1,7 @@
 mod barrier_cell;
 mod builder;
 mod channel_config;
+#[cfg(not(target_arch = "wasm32"))]
 mod direct_url_query;
 mod error;
 mod local_subdir;
@@ -13,7 +14,6 @@ mod subdir_builder;
 
 use std::{
     collections::HashSet,
-    path::PathBuf,
     sync::{Arc, Weak},
 };
 
@@ -23,13 +23,14 @@ pub use channel_config::{ChannelConfig, SourceConfig};
 use dashmap::{mapref::entry::Entry, DashMap};
 pub use error::GatewayError;
 pub use query::{NamesQuery, RepoDataQuery};
+#[cfg(not(target_arch = "wasm32"))]
 use rattler_cache::package_cache::PackageCache;
 use rattler_conda_types::{Channel, MatchSpec, Platform};
 pub use repo_data::RepoData;
 use reqwest_middleware::ClientWithMiddleware;
 use subdir::Subdir;
 use tokio::sync::broadcast;
-use tracing::instrument;
+use tracing::{instrument, Level};
 use url::Url;
 
 use crate::{gateway::subdir_builder::SubdirBuilder, Reporter};
@@ -158,9 +159,11 @@ struct GatewayInner {
     channel_config: ChannelConfig,
 
     /// The directory to store any cache
-    cache: PathBuf,
+    #[cfg(not(target_arch = "wasm32"))]
+    cache: std::path::PathBuf,
 
     /// The package cache, stored to reuse memory cache
+    #[cfg(not(target_arch = "wasm32"))]
     package_cache: PackageCache,
 
     /// A semaphore to limit the number of concurrent requests.
@@ -308,7 +311,7 @@ mod test {
     use rstest::rstest;
     use url::Url;
 
-    use crate::fetch::with_cache::CacheAction;
+    use crate::fetch::CacheAction;
     use crate::{
         gateway::Gateway,
         utils::{simple_channel_server::SimpleChannelServer, test::fetch_repo_data},
@@ -369,6 +372,7 @@ mod test {
     }
 
     #[tokio::test]
+    #[cfg(not(target_arch = "wasm32"))]
     async fn test_direct_url_spec_from_gateway() {
         let gateway = Gateway::builder()
             .with_package_cache(PackageCache::new(
