+++ crates/rattler_repodata_gateway/src/gateway/query.rs
@@ -9,7 +9,7 @@ use itertools::Itertools;
 use rattler_conda_types::{Channel, MatchSpec, Matches, PackageName, Platform};
 
 use super::{subdir::Subdir, BarrierCell, GatewayError, GatewayInner, RepoData};
-use crate::{gateway::direct_url_query::DirectUrlQuery, Reporter};
+use crate::Reporter;
 
 /// Represents a query to execute with a [`Gateway`].
 ///
@@ -171,37 +171,46 @@ impl RepoDataQuery {
         let mut pending_records = FuturesUnordered::new();
 
         // Push the direct url queries to the pending_records.
-        for (spec, url, name) in direct_url_specs {
-            let gateway = self.gateway.clone();
-            pending_records.push(
-                async move {
-                    let query = DirectUrlQuery::new(
-                        url.clone(),
-                        gateway.package_cache.clone(),
-                        gateway.client.clone(),
-                        spec.sha256,
-                    );
-
-                    let record = query
-                        .execute()
-                        .await
-                        .map_err(|e| GatewayError::DirectUrlQueryError(url.to_string(), e))?;
-
-                    // Check if record actually has the same name
-                    if let Some(record) = record.first() {
-                        if record.package_record.name != name {
-                            // Using as_source to get the closest to the retrieved input.
-                            return Err(GatewayError::UrlRecordNameMismatch(
-                                record.package_record.name.as_source().to_string(),
-                                name.as_source().to_string(),
-                            ));
-                        }
-                    }
-                    // Push the direct url in the first subdir result for channel priority logic.
-                    Ok((0, SourceSpecs::Input(vec![spec]), record))
+        cfg_if::cfg_if! {
+            if #[cfg(target_arch = "wasm32")] {
+                if let Some((_,first_url,_)) = direct_url_specs.into_iter().next() {
+                    // Direct url queries are not supported in wasm.
+                    return Err(GatewayError::DirectUrlQueryNotSupported(first_url.to_string()));
                 }
-                .boxed(),
-            );
+            } else {
+                for (spec, url, name) in direct_url_specs {
+                    let gateway = self.gateway.clone();
+                    pending_records.push(
+                        async move {
+                            let query = super::direct_url_query::DirectUrlQuery::new(
+                                url.clone(),
+                                gateway.package_cache.clone(),
+                                gateway.client.clone(),
+                                spec.sha256,
+                            );
+
+                            let record = query
+                                .execute()
+                                .await
+                                .map_err(|e| GatewayError::DirectUrlQueryError(url.to_string(), e))?;
+
+                            // Check if record actually has the same name
+                            if let Some(record) = record.first() {
+                                if record.package_record.name != name {
+                                    // Using as_source to get the closest to the retrieved input.
+                                    return Err(GatewayError::UrlRecordNameMismatch(
+                                        record.package_record.name.as_source().to_string(),
+                                        name.as_source().to_string(),
+                                    ));
+                                }
+                            }
+                            // Push the direct url in the first subdir result for channel priority logic.
+                            Ok((0, SourceSpecs::Input(vec![spec]), record))
+                        }
+                        .boxed(),
+                    );
+                }
+            }
         }
 
         let len = subdirs.len() + direct_url_offset;
@@ -230,7 +239,7 @@ impl RepoDataQuery {
                                 }
                             }
                         }
-                        .boxed(),
+                        .boxed_local(),
                     );
                 }
             }
@@ -308,10 +317,16 @@ impl RepoDataQuery {
 
 impl IntoFuture for RepoDataQuery {
     type Output = Result<Vec<RepoData>, GatewayError>;
+    #[cfg(not(target_arch = "wasm32"))]
     type IntoFuture = futures::future::BoxFuture<'static, Self::Output>;
+    #[cfg(target_arch = "wasm32")]
+    type IntoFuture = futures::future::LocalBoxFuture<'static, Self::Output>;
 
     fn into_future(self) -> Self::IntoFuture {
-        self.execute().boxed()
+        #[cfg(not(target_arch = "wasm32"))]
+        return self.execute().boxed();
+        #[cfg(target_arch = "wasm32")]
+        self.execute().boxed_local()
     }
 }
 
@@ -406,10 +421,16 @@ impl NamesQuery {
 
 impl IntoFuture for NamesQuery {
     type Output = Result<Vec<PackageName>, GatewayError>;
+    #[cfg(not(target_arch = "wasm32"))]
     type IntoFuture = futures::future::BoxFuture<'static, Self::Output>;
+    #[cfg(target_arch = "wasm32")]
+    type IntoFuture = futures::future::LocalBoxFuture<'static, Self::Output>;
 
     fn into_future(self) -> Self::IntoFuture {
-        self.execute().boxed()
+        #[cfg(not(target_arch = "wasm32"))]
+        return self.execute().boxed();
+        #[cfg(target_arch = "wasm32")]
+        self.execute().boxed_local()
     }
 }
