+++ crates/rattler_solve/Cargo.toml
@@ -31,7 +31,7 @@ criterion = { workspace = true }
 insta = { workspace = true, features = ["yaml"] }
 once_cell = { workspace = true }
 rattler_repodata_gateway = { path = "../rattler_repodata_gateway", default-features = false, features = [
-    "sparse",
+    "sparse", "memmap2"
 ] }
 rstest = { workspace = true }
 serde_json = { workspace = true }
