+++ js-rattler/Cargo.toml
@@ -31,7 +31,7 @@ wee_alloc = { version = "0.4.5", optional = true }
 thiserror = "2.0.3"
 
 rattler_conda_types = { path = "../crates/rattler_conda_types" }
-rattler_repodata_gateway = { path = "../crates/rattler_repodata_gateway" }
+rattler_repodata_gateway = { path = "../crates/rattler_repodata_gateway", features = ["gateway"] }
 
 # By adding the `libbz2-rs-sys` feature we ensure that bzip2 is using the rust
 # implementation. This is important because the C implementation is not
