+++ py-rattler/Cargo.toml
@@ -33,6 +33,7 @@ rattler = { path = "../crates/rattler", default-features = false, features = [
 rattler_repodata_gateway = { path = "../crates/rattler_repodata_gateway", default-features = false, features = [
     "sparse",
     "gateway",
+    "memmap2"
 ] }
 rattler_conda_types = { path = "../crates/rattler_conda_types", default-features = false }
 rattler_digest = { path = "../crates/rattler_digest" }
